requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pyodbc>=4.0.39
Pillow>=10.0.0
lxml>=4.9.0
fake-useragent>=1.4.0
python-dotenv>=1.0.0
urllib3>=2.0.0 